Adding OpenGL ES Support (#29)
* Adding OpenGL ES Support

* Ran CPPLINT

* Switching doubles to floats in inputMonitor

Changes:
* Input controller doesn't require the accuracy of doubles, so I switched them for floats. This could help embedded performance a bit.

* Removed double literals

---------

+++ src/main/engine/Misc/headers/inputMonitor.hpp
@@ -12,5 +12,5 @@
 #pragma once
 #include <game.hpp>
 void rotateShape(void *gameInfoStruct, void *target);
-double convertNegToDeg(double degree);
-double angleOfPoint(vec3 p1, vec3 p2);
+float convertNegToDeg(float degree);
+float angleOfPoint(vec3 p1, vec3 p2);
